# Models/Tblmusteri.cs
using System;
using System.Collections.Generic;

#nullable disable

namespace BulutBilisim.Models
{
    public partial class Tblmusteri
    {
        public int Musteriid { get; set; }
        public string Musteriad { get; set; }
        public string Musterisoyad { get; set; }
        public string Musteriemail { get; set; }
        public string Musterisifre { get; set; }
        public string Musteritelefon { get; set; }
    }
}