# Models/Tblurun.cs
using System;
using System.Collections.Generic;

#nullable disable

namespace BulutBilisim.Models
{
    public partial class Tblurun
    {
        public int Urunid { get; set; }
        public string Urunad { get; set; }
        public string Urunkategori { get; set; }
        public string Urunmarka { get; set; }
        public string Urunfiyat { get; set; }
        public string Urunfoto1 { get; set; }
        public string Urunfoto2 { get; set; }
        public string Urunfoto3 { get; set; }
    }
}